requirements.prd.txt
Django
MySQL-python
PyYAML
django-model-utils
django-extensions
django-pyc
djangorestframework
markdown
django-filter
